Removed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2265 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/projectile_manager.cpp
@@ -163,7 +163,7 @@ void ProjectileManager::updateClient(float dt)
     unsigned int num_projectiles = race_state->getNumFlyables();
     if(num_projectiles != m_active_projectiles.size())
         fprintf(stderr, "Warning: num_projectiles %d active %d\n",num_projectiles,
-                m_active_projectiles.size());
+                (int)m_active_projectiles.size());
 
     unsigned int indx=0;
     for(Projectiles::iterator i  = m_active_projectiles.begin();
